+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New filters design on macOS
 - Add missing "Remove Pins" button on macOS
 - Fix an issue where using "Recent" date filter was applying "Session" filter instead
+- Fix an issue where when you were focusing a log label, it wasn't reflected in the filters
 
 ## Pulse 3.0.0